+++ bindings/nostr-sdk-js/examples/negentropy.js
@@ -1,4 +1,4 @@
-const { loadWasmAsync, initLogger, LogLevel, NegentropyOptions, NegentropyDirection, Filter, ClientBuilder, NostrDatabase } = require("../");
+const { loadWasmAsync, initLogger, LogLevel, NegentropyOptions, NegentropyDirection, Filter, Client, NostrDatabase } = require("../");
 
 // NOTE: this code work only on browser (due to indexeddb)!
 
@@ -7,8 +7,8 @@ async function main() {
 
     initLogger(LogLevel.info());
 
-    let db = NostrDatabase.indexeddb("js-test");
-    let client = new ClientBuilder().database(db).build();
+    let db = await NostrDatabase.indexeddb("js-test");
+    let client = Client.builder().database(db).build();
 
     await client.addRelay("wss://relay.damus.io");